requirements.txt
# Core PDF parsing
pdfplumber>=0.10.0
PyMuPDF>=1.23.0

# Web scraping
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
googlesearch-python>=1.2.0

# Text matching & ML
rapidfuzz>=3.5.0
scikit-learn>=1.3.0

# Templating & config
Jinja2>=3.1.0
PyYAML>=6.0.0

# Web application
Flask>=3.0.0

# LLM integration (local Ollama)
ollama>=0.4.0
